robot/requirements.txt
robotframework
robotframework-requests
robotframework-seleniumlibrary
robotframework-faker
